# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * National Empowerment Scheme - Training and Vocational Skills Registration Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Applicant {
  id: number;
  registrationNumber: string;
  fullName: string;
  gender: string;
  dateOfBirth: string;
  phoneNumber: string;
  email: string;
  residentialAddress: string;
  state: string;
  lga: string;
  stateOfOrigin: string;
  lgaOfOrigin: string;
  nin: string;
  bankName: string;
  bankAccountNumber: string;
  /** @nullable */
  passportPhotoUrl?: string | null;
  highestEducation: string;
  schoolAttended: string;
  graduationYear: string;
  skills: string[];
  /** @nullable */
  otherSkill?: string | null;
  hasPreviousExperience: boolean;
  /** @nullable */
  previousSkillDescription?: string | null;
  trainingExpectation: string;
  employmentStatus: string;
  hasExistingBusiness: boolean;
  /** @nullable */
  businessDescription?: string | null;
  declarationName: string;
  /** @nullable */
  digitalSignature?: string | null;
  declarationDate: string;
  agreedToTerms: boolean;
  status: string;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ApplicantInput {
  fullName: string;
  gender: string;
  dateOfBirth: string;
  phoneNumber: string;
  email: string;
  residentialAddress: string;
  state: string;
  lga: string;
  stateOfOrigin: string;
  lgaOfOrigin: string;
  nin: string;
  bankName: string;
  bankAccountNumber: string;
  passportPhotoUrl?: string;
  highestEducation: string;
  schoolAttended: string;
  graduationYear: string;
  skills: string[];
  otherSkill?: string;
  hasPreviousExperience: boolean;
  previousSkillDescription?: string;
  trainingExpectation: string;
  employmentStatus: string;
  hasExistingBusiness: boolean;
  businessDescription?: string;
  declarationName: string;
  digitalSignature?: string;
  declarationDate: string;
  agreedToTerms: boolean;
}

export interface DraftInput {
  fullName?: string;
  gender?: string;
  dateOfBirth?: string;
  phoneNumber?: string;
  email?: string;
  residentialAddress?: string;
  state?: string;
  lga?: string;
  stateOfOrigin?: string;
  lgaOfOrigin?: string;
  nin?: string;
  bankName?: string;
  bankAccountNumber?: string;
  passportPhotoUrl?: string;
  highestEducation?: string;
  schoolAttended?: string;
  graduationYear?: string;
  skills?: string[];
  otherSkill?: string;
  hasPreviousExperience?: boolean;
  previousSkillDescription?: string;
  trainingExpectation?: string;
  employmentStatus?: string;
  hasExistingBusiness?: boolean;
  businessDescription?: string;
  declarationName?: string;
  digitalSignature?: string;
  declarationDate?: string;
  agreedToTerms?: boolean;
}

export interface ApplicantUpdate {
  status?: string;
  fullName?: string;
  phoneNumber?: string;
  email?: string;
}

export interface ApplicantListResponse {
  data: Applicant[];
  total: number;
  page: number;
  limit: number;
}

export interface StatEntry {
  label: string;
  count: number;
}

export interface ApplicantStats {
  totalRegistrations: number;
  maleCount: number;
  femaleCount: number;
  otherGenderCount: number;
  draftCount: number;
  submittedCount: number;
  byState: StatEntry[];
  bySkill: StatEntry[];
  byEducation: StatEntry[];
  byEmploymentStatus: StatEntry[];
  recentRegistrations: Applicant[];
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface Admin {
  id: number;
  username: string;
  role: string;
}

export interface AuthResponse {
  success: boolean;
  admin: Admin;
}

export type ListApplicantsParams = {
/**
 * Search by name, email, or registration number
 */
search?: string;
/**
 * Filter by state
 */
state?: string;
/**
 * Filter by skill category
 */
skill?: string;
/**
 * Filter by gender
 */
gender?: string;
/**
 * Filter by status (submitted, draft)
 */
status?: string;
/**
 * Page number
 */
page?: number;
/**
 * Items per page
 */
limit?: number;
};