# lib/api-zod/src/generated/types/applicant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * National Empowerment Scheme - Training and Vocational Skills Registration Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Applicant {
  id: number;
  registrationNumber: string;
  fullName: string;
  gender: string;
  dateOfBirth: string;
  phoneNumber: string;
  email: string;
  residentialAddress: string;
  state: string;
  lga: string;
  stateOfOrigin: string;
  lgaOfOrigin: string;
  nin: string;
  bankName: string;
  bankAccountNumber: string;
  /** @nullable */
  passportPhotoUrl?: string | null;
  highestEducation: string;
  schoolAttended: string;
  graduationYear: string;
  skills: string[];
  /** @nullable */
  otherSkill?: string | null;
  hasPreviousExperience: boolean;
  /** @nullable */
  previousSkillDescription?: string | null;
  trainingExpectation: string;
  employmentStatus: string;
  hasExistingBusiness: boolean;
  /** @nullable */
  businessDescription?: string | null;
  declarationName: string;
  /** @nullable */
  digitalSignature?: string | null;
  declarationDate: string;
  agreedToTerms: boolean;
  status: string;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}