# lib/api-zod/src/generated/types/applicantListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * National Empowerment Scheme - Training and Vocational Skills Registration Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Applicant } from './applicant';

export interface ApplicantListResponse {
  data: Applicant[];
  total: number;
  page: number;
  limit: number;
}