# lib/api-zod/src/generated/types/applicantStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * National Empowerment Scheme - Training and Vocational Skills Registration Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Applicant } from './applicant';
import type { StatEntry } from './statEntry';

export interface ApplicantStats {
  totalRegistrations: number;
  maleCount: number;
  femaleCount: number;
  otherGenderCount: number;
  draftCount: number;
  submittedCount: number;
  byState: StatEntry[];
  bySkill: StatEntry[];
  byEducation: StatEntry[];
  byEmploymentStatus: StatEntry[];
  recentRegistrations: Applicant[];
}