# lib/api-zod/src/generated/types/applicantUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * National Empowerment Scheme - Training and Vocational Skills Registration Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicantUpdate {
  status?: string;
  fullName?: string;
  phoneNumber?: string;
  email?: string;
}