# lib/api-zod/src/generated/types/draftInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * National Empowerment Scheme - Training and Vocational Skills Registration Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface DraftInput {
  fullName?: string;
  gender?: string;
  dateOfBirth?: string;
  phoneNumber?: string;
  email?: string;
  residentialAddress?: string;
  state?: string;
  lga?: string;
  stateOfOrigin?: string;
  lgaOfOrigin?: string;
  nin?: string;
  bankName?: string;
  bankAccountNumber?: string;
  passportPhotoUrl?: string;
  highestEducation?: string;
  schoolAttended?: string;
  graduationYear?: string;
  skills?: string[];
  otherSkill?: string;
  hasPreviousExperience?: boolean;
  previousSkillDescription?: string;
  trainingExpectation?: string;
  employmentStatus?: string;
  hasExistingBusiness?: boolean;
  businessDescription?: string;
  declarationName?: string;
  digitalSignature?: string;
  declarationDate?: string;
  agreedToTerms?: boolean;
}